requirements.txt
farm-haystack==1.5.0
s3fs==2022.5.0
ray==1.13.0
ray[serve]==1.13.0
httpx==0.23.0
